Change new window in Edit Profile to Bootstrap modal
The way that profile editing was implented was going to break backend logic. Having the editing pane open in a separate browser window completely disconnects it from the original window.

Using a modal is the proper way to do what was intended here. Just copied over elements from the original popout and made a couple class adjustments to allow for the POST request to be handled.